HADOOP-15820. ZStandardDecompressor native code sets an integer field as a long. Contributed by Jason Lowe
(cherry picked from commit f13e231025333ebf80b30bbdce1296cef554943b)

+++ hadoop-common-project/hadoop-common/src/main/native/src/org/apache/hadoop/io/compress/zstd/ZStandardDecompressor.c
@@ -145,7 +145,7 @@ JNIEXPORT void JNICALL Java_org_apache_hadoop_io_compress_zstd_ZStandardDecompre
         THROW(env, "java/lang/InternalError", dlsym_ZSTD_getErrorName(result));
         return;
     }
-    (*env)->SetLongField(env, this, ZStandardDecompressor_remaining, 0);
+    (*env)->SetIntField(env, this, ZStandardDecompressor_remaining, 0);
 }
